+++ src/imio/helpers/tests/test_cache.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 
 from imio.helpers.cache import cleanVocabularyCacheFor
+from imio.helpers.cache import cleanRamCache
 from imio.helpers.cache import cleanRamCacheFor
 from imio.helpers.testing import IntegrationTestCase
 from zope.component import queryUtility
@@ -68,6 +69,23 @@ class TestCacheModule(IntegrationTestCase):
         cleanVocabularyCacheFor()
         self.assertEquals([term.token for term in vocab(self.portal)._terms],
                           ['1', '2', '3', '4', '5'])
+
+    def test_cleanRamCache(self):
+        """
+        This helper method invalidates all ram.cache.
+        """
+        self.portal.REQUEST.set('ramcached', 'a')
+        self.assertEquals(ramCachedMethod(self.portal, param='1'), 'a')
+        # change value in REQUEST, as it is ram cached, it will still return 'a'
+        self.portal.REQUEST.set('ramcached', 'b')
+        self.assertEquals(ramCachedMethod(self.portal, param='1'), 'a')
+        # ram.cache works as expected if param changes
+        self.assertEquals(ramCachedMethod(self.portal, param='2'), 'b')
+        # try again
+        self.assertEquals(ramCachedMethod(self.portal, param='1'), 'a')
+        # now clean all caches, it will returns 'b'
+        cleanRamCache()
+        self.assertEquals(ramCachedMethod(self.portal, param='1'), 'b')
 
     def test_cleanRamCacheFor(self):
         """